requirements.txt
Flask==2.1.1
requests==2.27.1
snowflake-connector-python==2.7.2
gunicorn==20.1.0
